feat(media): refactor upgrades section with new content structure and improved functionality
- Introduced UpgradesPageContent component to streamline the upgrades page layout and data fetching.
- Updated UpgradesPage to utilize the new component, enhancing readability and maintainability.
- Implemented initialSelectedSlug handling for better navigation to specific upgrade details.
- Enhanced the SIMDDetailPanel to support related SIMD selection, improving user interaction.
- Adjusted styles and layout for better user experience across the upgrades section.
