+++ apps/media/app/[locale]/upgrades/[slug]/page.tsx
@@ -1,0 +1,19 @@
+import type { Metadata } from "next";
+import { UpgradesPageContent } from "../upgrades-page-content";
+import { upgradesListingMetadata } from "@/lib/metadata";
+
+export const revalidate = 300;
+
+export async function generateMetadata(): Promise<Metadata> {
+  return upgradesListingMetadata();
+}
+
+export default async function UpgradeDetailPage({
+  params,
+}: {
+  params: Promise<{ locale: string; slug: string }>;
+}) {
+  const { locale, slug } = await params;
+
+  return <UpgradesPageContent locale={locale} initialSelectedSlug={slug} />;
+}
